fix(lint): home the nil-receiver tests where both size gates hold
Folding the tests into event_history_test.go traded one lint failure for two:
that file reached 1007 lines against a 1000 limit, which backend-size and
pkg-maint (rule=file-lines) both report.

The package cannot take a new file (pkg-file-count baseline 19) and this file
cannot grow, so put the tests in test_history_test.go, which had 40 lines of
headroom and is already the home of the history's construction-contract tests
("NewFactoryEventHistory rejects missing clock"). A nil-receiver test is the
exact complement of those: what happens when a caller uses a history that was
never constructed.

Verified locally, all three gates together:
- backend-size: LINT_VIOLATION_COUNT 2, equal to the recorded baseline
- pkg-maint: passed (file lines <= 1000, function lines <= 100, complexity <= 15)
- pkg-file-count: holds, package still at 19 files
go vet clean and both nil-receiver tests pass.
